+++ api/src/main/java/org/grnet/cat/api/endpoints/AssessmentsV2Endpoint.java
@@ -34,7 +34,11 @@ import org.grnet.cat.constraints.NotFoundEntity;
 import org.grnet.cat.dtos.InformativeResponse;
 import org.grnet.cat.dtos.assessment.registry.JsonRegistryAssessmentRequest;
 import org.grnet.cat.dtos.assessment.registry.UserJsonRegistryAssessmentResponse;
+import org.grnet.cat.repositories.ActorRepository;
+import org.grnet.cat.repositories.AssessmentTypeRepository;
 import org.grnet.cat.repositories.MotivationAssessmentRepository;
+import org.grnet.cat.repositories.registry.MotivationRepository;
+import org.grnet.cat.repositories.registry.RegistryActorRepository;
 import org.grnet.cat.services.assessment.JsonAssessmentService;
 import org.grnet.cat.utils.Utility;
 
@@ -338,5 +342,60 @@ public class AssessmentsV2Endpoint {
         var assessment = assessmentService.update(id, request);
 
         return Response.ok().entity(assessment).build();
+    }
+
+    @Tag(name = "Assessment")
+    @Operation(
+            summary = "Get list of public assessment objects created / used by a specific user by type and actor.",
+            description = "This endpoint is public and any unauthenticated user can retrieve published assessment objects categorized by motivation and actor, created by all users." +
+                    "By default, the first page of 10 public assessment objects will be returned. You can tune the default values by using the query parameters page and size.")
+    @APIResponse(
+            responseCode = "200",
+            description = "List of public assessment objects.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = AssessmentsEndpoint.PageableObjects.class)))
+    @APIResponse(
+            responseCode = "400",
+            description = "Bad Request",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "404",
+            description = "Entity Not Found.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "500",
+            description = "Internal Server Error.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @GET
+    @Path("/public-objects/by-motivation/{motivation-id}/by-actor/{actor-id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response objectsByTypeAndActor(@Parameter(
+            description = "The Motivation.",
+            required = true,
+            example = "pid_graph:3E109BBA",
+            schema = @Schema(type = SchemaType.STRING))
+                                          @PathParam("motivation-id") @Valid @NotFoundEntity(repository = MotivationRepository.class, message = "There is no Motivation with the following id:") String motivationId, @Parameter(
+            description = "The Actor to retrieve public assessment objects.",
+            required = true,
+            example = "pid_graph:B5CC396B",
+            schema = @Schema(type = SchemaType.STRING))
+                                          @PathParam("actor-id") @Valid @NotFoundEntity(repository = RegistryActorRepository.class, message = "There is no Actor with the following id:") String actorId,
+                                          @Parameter(name = "page", in = QUERY,
+                                                  description = "Indicates the page number. Page number must be >= 1.") @DefaultValue("1") @Min(value = 1, message = "Page number must be >= 1.") @QueryParam("page") int page,
+                                          @Parameter(name = "size", in = QUERY,
+                                                  description = "The page size.") @DefaultValue("10") @Min(value = 1, message = "Page size must be between 1 and 100.")
+                                          @Max(value = 100, message = "Page size must be between 1 and 100.") @QueryParam("size") int size,
+                                          @Context UriInfo uriInfo) {
+
+        var assessments = assessmentService.getPublishedAssessmentObjectsByMotivationAndActorAndPage(page - 1, size, motivationId, actorId, uriInfo);
+
+        return Response.ok().entity(assessments).build();
     }
 }